qr and object module refactor
